Cleaning up some more recent g++ compiler errors.
One of them appears not to make sense.

+++ grail/lib/io/CommandLineOptions.cpp
@@ -482,7 +482,7 @@ namespace grail { namespace io {
                 "note",
                 FLTL_F_BLUE,
                 diag,
-                argv_offset,
+                static_cast<int>(argv_offset),
                 offset
             );
         }
